major backend changes and react frontend with all user and admin functionality available

+++ src/main/java/com/eventmanagement/backend/repository/VenueRepository.java
@@ -3,5 +3,5 @@ package com.eventmanagement.backend.repository;
 import com.eventmanagement.backend.entity.Venues;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface VenueRepository extends JpaRepository<Venues,Integer> {
+public interface VenueRepository extends JpaRepository<Venues,Long> {
 }
